% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions_data.R
\docType{data}
\name{assumptions_data}
\alias{assumptions_data}
\title{Data set for practicing model assumptions (synthetic).}
\format{
A data frame with 50 rows and 3 variables:
\describe{
\item{x}{continuous variable}
\item{y}{continuous variable}
\item{y_quad}{y with a quadratic trend}
}
}
\source{
\url{https://github.com/jvcasillas/ds4ling}
}
\usage{
assumptions_data
}
\description{
A synthetic dataset for teaching model scrutiny
}
\keyword{datasets}
