% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumption.R
\docType{data}
\name{consumption}
\alias{consumption}
\title{Food consumption data set for practicing linear regression (synthetic).
Taken from QASS 57, Ch. 1, table 1}
\format{
A data frame with 50 rows and 4 columns:
\describe{
\item{food_consumption}{amount spent in dollars on food}
\item{income}{annual income in dollars}
\item{family_size}{number of family members in household}
\item{live_on_farm}{y/n regarding whether or not family lives on a farm}
}
}
\source{
\url{https://github.com/jvcasillas/ds4ling}
}
\usage{
consumption
}
\description{
A synthetic dataset for teaching the linear model
}
\keyword{datasets}
