% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{diagnosis}
\alias{diagnosis}
\title{A function to plot model diagnostics}
\usage{
diagnosis(mod)
}
\arguments{
\item{mod}{A model object}

\item{diagnosis}{A function for plotting model diagnostics}
}
\description{
This function generates a series of plots that can be used to check some
model assumptions in linear models.
}
\details{
The left panel shows residuals as a function of fitted values and is useful
for checking homoskedasticity.

The middle panel plots a histogram of the residuals with a vertical
discontinuous line that represents the mean.

The right panel provides a qq-plot of the residuals.
}
\examples{
mod1 <- lm(mpg ~ drat, data = mtcars)
diagnosis(mod1)
}
\keyword{plot}
