% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds4ling_theme.R
\name{ds4ling_theme}
\alias{ds4ling_theme}
\title{ds4ling plotting theme}
\usage{
ds4ling_theme(...)
}
\arguments{
\item{...}{Pass extra parameters}

\item{ds4ling_theme}{A theme function used with ggplot (e.g., 'theme_bw')}
}
\description{
This function provides the plot theme used in most ds4ling slides
}
\examples{
library(ggplot2)
library(untidydata)
ggplot(pre_post, aes(x = test1, y = test2)) +
  geom_point() +
  ds4ling_theme()

}
\keyword{plot}
