% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_qqplot.R
\name{gg_qqplot}
\alias{gg_qqplot}
\title{QQ-plot function}
\usage{
gg_qqplot(vec)
}
\arguments{
\item{vec}{A numeric vector}

\item{gg_qqplot}{A function for generating QQ-plots}
}
\description{
This function creates a QQ-plot using ggplot
}
\examples{
gg_qqplot(rnorm(30))

}
\keyword{plot}
