% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice_cream_poisson_data.R
\docType{data}
\name{ice_cream_poisson_data}
\alias{ice_cream_poisson_data}
\title{Coint data for teaching poisson regression (synthetic).}
\format{
A data frame with 2600 rows and 2 variables:
\describe{
\item{temp}{Temperature in F}
\item{city}{City of ice cream consumption}
\item{units}{Counts of ice creams sold}
}
}
\source{
\url{https://github.com/jvcasillas/ds4ling}
}
\usage{
ice_cream_poisson_data
}
\description{
A synthetic dataset of ice cream consumption.
}
\keyword{datasets}
