% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrc_ex_data.R
\docType{data}
\name{mrc_ex_data}
\alias{mrc_ex_data}
\title{Age, height, and math score data (synthetic).}
\format{
A data frame with 10000 rows and 5 variables:
\describe{
\item{age}{Age of participants}
\item{height}{Height of participants}
\item{score}{Math proficiency score}
\item{age_c}{Mean centered age}
\item{height_c}{Mean centered height}
}
}
\source{
\url{https://github.com/jvcasillas/ds4ling}
}
\usage{
mrc_ex_data
}
\description{
A synthetic dataset for teaching MRC.
}
\keyword{datasets}
