% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{A function to plot distributions}
\usage{
plot_dist(distribution = "gaussian", range = c(-3, 3), mean = 0, sd = 1, ...)
}
\arguments{
\item{distribution}{A distribution type (defaults to gaussian)}

\item{range}{A vector of 2 values indicating the range for the x-axis and
the data}

\item{mean}{A numeric value for the center of gravity of the distribution
(defaults to 0)}

\item{sd}{A numeric value for the standard deviation of the distribution
(defaults to 1)}

\item{...}{Arguments to be passed to ggplot. Specifically 'font_color' and
'tex'.}

\item{plot_dist}{A function for plotting normal distributions}
}
\description{
This function generates a plot of a normal distribution using user
supplied mean and standard deviation.
}
\examples{
plot_dist(font_color = "darkblue")
plot_dist(range = c(-10, 10), distribution = "gaussian", mean = 1, sd = 3)
plot_dist(range = c(-1, 1), distribution = "gaussian", mean = 0, sd = 0.3)
plot_dist(range = c(-5, 5), distribution = "anything")
plot_dist(tex = "$Hello_{\\\\Delta} = 4^5$")
}
\keyword{plot}
