% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverside.R
\docType{data}
\name{riverside}
\alias{riverside}
\title{Riverside data set for practicing linear regression (synthetic).
Taken from QASS 22, Ch. 1, table 2}
\format{
A data frame with 32 rows and 3 columns:
\describe{
\item{respondent}{participant id variable}
\item{education}{education in years}
\item{income}{income in dollars}
}
}
\source{
\url{https://github.com/jvcasillas/ds4ling}
}
\usage{
riverside
}
\description{
A synthetic dataset for teaching the linear model
}
\keyword{datasets}
