% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_scores.R
\docType{data}
\name{test_scores}
\alias{test_scores}
\title{Test score data (synthetic).}
\format{
A data frame with 4000 rows and 4 variables:
\describe{
\item{id}{Participant ID}
\item{group}{Categorical variable denoting which class the student is in}
\item{score}{Raw test score}
\item{important}{Indicator variable for 'you' and your 'friend'}
}
}
\source{
\url{https://github.com/jvcasillas/ds4ling}
}
\usage{
test_scores
}
\description{
A synthetic dataset for teaching z-transformations
}
\keyword{datasets}
