% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_scores_rm.R
\docType{data}
\name{test_scores_rm}
\alias{test_scores_rm}
\title{Repeated measures test score data (synthetic).}
\format{
A data frame with 16 rows and 4 variables:
\describe{
\item{id}{Participant ID}
\item{spec}{Grouping variable (untidy) specifying group number and level}
\item{test1}{Raw pre-test score}
\item{test2}{Raw post-test score}
}
}
\source{
\url{https://github.com/jvcasillas/ds4ling}
}
\usage{
test_scores_rm
}
\description{
A synthetic dataset for teaching correlations
}
\keyword{datasets}
