% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab_data.R
\docType{data}
\name{vocab_data}
\alias{vocab_data}
\title{Vocabulary data (synthetic).}
\format{
A data frame with 4000 rows and 4 variables:
\describe{
\item{ages}{Age of participants}
\item{vocab}{Estimated vocabulary size}
\item{count}{Row numbers}
\item{reader_type}{How often the participant reads (average, frequent)}
}
}
\source{
\url{https://github.com/jvcasillas/ds4ling}
}
\usage{
vocab_data
}
\description{
A synthetic dataset for teaching correlation.
}
\keyword{datasets}
