% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vot_logistic_data.R
\docType{data}
\name{vot_logistic_data}
\alias{vot_logistic_data}
\title{VOT perception data for logistic regression (synthetic).}
\format{
A data frame with 2600 rows and 2 variables:
\describe{
\item{resp}{Binary response (/b/-/p/)}
\item{vot}{Continuum step}
}
}
\source{
\url{https://github.com/jvcasillas/ds4ling}
}
\usage{
vot_logistic_data
}
\description{
A synthetic dataset for teaching logistic regression.
}
\keyword{datasets}
